requirements.txt
PyJWT>=2.8.0
fastapi>=0.100.0
httpx>=0.25.0
pydantic>=1.10.0
pytest>=7.0.0
uvicorn>=0.20.0
